+++ src/main/java/com/tenpo/adder/addition/service/AdditionService.java
@@ -2,6 +2,6 @@ package com.tenpo.adder.addition.service;
 
 public interface AdditionService {
 
-    Long calculateAddition(Long numberOne, Long numberTwo);
+    Integer calculateAddition(Integer numberOne, Integer numberTwo);
 
 }
