+++ src/main/java/com/tenpo/adder/addition/service/AdditionServiceImpl.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Service;
 public class AdditionServiceImpl implements AdditionService{
 
     @Override
-    public Long calculateAddition(Long numberOne, Long numberTwo) {
+    public Integer calculateAddition(Integer numberOne, Integer numberTwo) {
         return Math.addExact(numberOne, numberTwo);
     }
 }
